Boot fix and RRAM size change
This commit fixes an issue that prevented the bootloader to load the
stack pointer properly when booting the image. The previous code:

```
         // Configure image stack pointer
         __set_MSP( *image_base );

         enter_image();
```

resulted in this assembler output:

```
...
        msr MSP, r3 // SP loaded here
...
        add sp, #8  // oops!
        bx  r2      // jump to image
```

Notice that GCC tries to restore the stack before jumping into the image
(`add sp, #8`) which is the expected behavior, but since SP is already set
at the end of RAM (by the firmware image) in `msr MSP, r3`, this will set
SP to a region outside the available RAM, which in turn will trigger a
hard fault right when the firmware image runs (no idea how this ever
worked).

With the new code in place, the assembler output looks like this:

```
...
        msr MSP, r3      // SP loaded here
...
        ldr r3, [r3, #0] // load image address
        mov pc, r3       // jump to image immediately
        add sp, #8       // dead code
        bx  lr           // dead code
```

GCC still tries to clear the stack, but the `add sp, #8` instruction is
never reached because of the previous `mov pc, r3`.

The commit also changes the RRAM size from 32 bytes to 128 bytes. A
separate commit will be needed to the eLua repository to enable this;
also, we need to make sure that the 128-bytes RRAM bootloader is used
only with 128-bytes RRAM firmware images and the other way around
(although a 32-byte RRAM firmware image should also work with an
128-bytes RRAM bootloader, it's probably best to mix and match them
properly to avoid confusion).